Model event metadata
The reason behind the UnpersistedMetadata struct is to represent the
metadata of an unpersisted event. For instance, we don't know the
create_at of an event that hasn't been created yet.

We could make some of the metadata fields optional or omittable but that
would result in our persisted events also having optional/omittable
fields which could lead to unexpected data.