Major revision of README
I’ve re-written the README file.

Specifically, I wanted to describe the flow of information (Command -> Handler -> Event) and how the pieces of our system address that flow.

I’m also using it as a means to propose some minor revisions to the API we’re exposing; specifically how `metadata` comes in via the `CommandHandler`, through the `Aggregate` and ultimately ends up in the `Event`

Finally, I want to use this as a starting-point for discussion about our structure and API with a few other campers. Will discuss more on Slack.