Refactoring Aggregate / AggregateSpec
Three big changes:

1. Event handlers can now receive metadata by including a second block argument

2. The example domain events (and Aggregate) have been re-built to better reflect the new domain-module structure, as well as to demonstrate the metadata-capture feature

3. `Aggregate.load_from_history` has been replaced with `Aggregate#load_history`

`load_from_history` was doing some awkward backflips in order to allow sub-classes to use an initialiser to set up internal state. On further reflection, I felt like this was a bit of a smell—any initial internal state the aggregate needs to be complete should be managed by the first event in the sequence.

This change makes  instantiation _much_ simpler, and removes the call to `instance_variable_set`