Implement sink_event / stage_event in Aggregate
BIG CHANGE ALERT.

The spec should be fairly descriptive—the biggest change in this is when we generate the the `aggregate_sequence` to be persisted in the database.

Currently, it’s being generated by the Event Sink; I would argue that the sink should simply try to persist whatever events we give it, and that the object calling `sink` should determine the `aggregate_sequence`—tell, don’t ask.

This happens in `generate_staged_event`.

The most controversial part of change is probably the `stage_event` method. It presumes that by staging an event, we want the Aggregate to believe that the event has been successfully sunk—including having its internal `aggregate_sequence` being updated.

If we stage a set of events and they fail to persist, the failure will bubble up to the CommandHandler, which should then re-run the entire process, including re-building the aggregate from history.

Any advice on this approach would be most welcome.