Limit class scope for DomainEvent deserialisation
There was an inadvertent bug in our framework spec that let our spec pass even though our domain objects weren’t defined in the correct module.

This fixes that problem, and prvents the (unlikely, but possible) situation where we have a class with the same name as one of our domain events outside of the `Domains` module.