CommandHandler retries command on failure
If `CommandHandler.callable` raises a `RetriableException`, the command handler will continually execute the block until it passes, or the retry-count crosses the threshold. If it crosses the threshold, a `RetryFailureThresholdExceededException` will be raised.

Currently, the threshold is 3; I picked it because 3 is a nice number. Open to suggestions for changing. If need be, we can overload the constant on a per-handler basis if we feel the need to.