Ensure graceful failure on missing EventStore DB
In order to give farm owners some lead-time in updating their farms,
make sure we fail gracefully if ParameterStore access or content isn't
properly configured.

This means running `bin/event_framework event_store:db:migrate` is
effectively a no-op if ParameterStore is misconfigured, meaning
we can safely run it on deploy.

The only path to failure on this should then be if someone sets
a feature flag that opens up a path to executing commands.