Remove "event type" scoping from event processors
We're doing this because we've seen cases where events have been skipped
by our event processors. We added logging and found that the query to
find the events skipped an event in the middle.

Example:

    last_sequence = 100
    get_after(last_sequence)
    # => [event_101, event_102, event_104] (event_103 is missing)

When we re-ran the same query (to the best of our knowledge) the event
would be included. We're not 100% sure this is the issue but simplifying
this query should rule it out at the very least.

We're also updating the event processor monitor so the numbers match up.