Isolate test database for EventFramework
The first thing we need to do in eliminating all references to
`EventStore.database` is to provide a databsae connection for the
EventFramework specs to use.

I wanted to ensure that this database was completely isolated from the
databases used for actual implementation, so I moved the setup and
configuration into a spec helper, created an independent set of
migrations, and used a different ENV variable.

There's no performance hit in doing the set-up in the `before(:suite)`
block.