BookmarkRepository handles multiple bookmarks
Continuing the pattern of passing _instances_ as dependencies, not
classes, I reasoned that it would make more sense to pass a configured
instance of BookmarkRepository that could be used multiple times,
rather than having to create _n_ instances of the injected class.

To make this happen, I simply added `name` as an argument to `checkout`

This also closes the loop on eliminating `BookmarkReadonly`, getting rid
of the need to manually instantiate `Bookmark` instances, pulling them
instead from the BookmarkRepository.