Switch to pg_try_advisory_xact_lock for event sink locking
Using the _xact_ version of the function allows Postgres to
automatically release the lock at the end of the transaction. This
leaves less chance for something to go wrong in the Ruby code and
prevent the lock getting release, which would prevent all event writes.

@melm73 had an issue in production with similar code where an insert
raised an exception and then the unlock raised a
`PG::InFailedSqlTransaction` error causing the lock to never be
released.

From the docs:

    pg_try_advisory_xact_lock works the same as pg_try_advisory_lock,
    except the lock, if acquired, is automatically released at the end
    of the current transaction and cannot be released explicitly.

    - https://www.postgresql.org/docs/9.1/functions-admin.html