Avoid persisting Metadata#migrated if false
Currently, we're persisting `migrated` and `bypass_mongo_projection`
in the metadata of _all_ our events, regardless of their value. This
is because we're specifying defaults in the metadata struct.

It'd be great if we didn't have to carry this forward into the
event-store rebuild. This change solves that problem.

This changes `Event#metadata.migrated` to return nil instead of false,
but given the way it is used in the after_sink_hook, this
won't cause us any grief.