Add EventStore integration test
This is a bit of a weird test, under the hood we're actually testing
that the EventBuilder can load both types of metadata. The EventBuilder
tests are unit tests and we can accidentally use the wrong test data by
simply passing in the wrong types.

I've decided to add a test at this level so that we have a solid
integration test on sinking and loading events.