Reinstate event processor logic to filter events by relevant types
This is effectively a revert of commit c45d5d3 but since a lot of the code has
changed since then a bunch of manual fixes needed to be made. Since we found
out that the event-skipping issue that this commit was trying to fix was solved
by locking the events table when sinking events (48cd92b) we can bring back
this optimization which should result in some performance gains when a large
number of events are sunk into the event stream at the same time and need to be
processed quickly.