Pass event_builder in as a dependency
If an event that is being processed does not conform to the specific
structure that the EventBuilder is expecting, the event store will fail.

By passing in the EventBuilder as a dependency, we are able to build an
Event Framework Event from any stored event structure.

Interestingly, ruby allows you to reference a given kwarg as a parameter
in a different default kwarg value, as in
```
event_builder: EventBuilder.new(event_type_resolver: event_type_resolver)
```