Use a blocking lock function in event store sink
Previously we were using pg_try_advisory_xact_lock to acquire a lock in
the event sink. This function does not block and therefore does not
guarantee that you've acquired the lock. To handle this we would try
getting the lock in a loop until we either succeeded or after a number
of retries we would raise a ConcurrencyError.

Instead of this we're switching to pg_advisory_xact_lock which does
block. This means that PostgreSQL will handle waiting for the lock.

Previously we were retrying 100 times and sleeping each time for 0.1
seconds. To match this we're setting our lock timeout to 10000
milliseconds.

Some basic testing using ./bin/demonstrate_event_sequence_id_gaps shows
about a 3x performance increase.